requirements.txt
aiohttp==0.20.2
chardet==2.3.0
pyserial==3.0.1
requests==2.9.1
telepot==6.2
wheel==0.24.0
